Moved existing functionality to separate class
